仮のAPIで認証とできなかった時のリダイレクト
