+++ frontend/src/app/client/[shop_id]/page.tsx
