+++ frontend/src/app/client/[shop_id]/waitList/[id]/page.tsx
