+++ frontend/src/lib/zzzoldnextauth.ts
@@ -1,0 +1,78 @@
+import { PrismaAdapter } from '@next-auth/prisma-adapter';
+import { getServerSession, type NextAuthOptions } from 'next-auth';
+import prisma from '@/lib/prisma';
+import GoogleProvider from 'next-auth/providers/google';
+import CredentialsProvider from 'next-auth/providers/credentials';
+import bcrypt from 'bcrypt';
+
+// NextAuth設定
+export const authOptions: NextAuthOptions = {
+  // Prismaを使うための設定
+  adapter: PrismaAdapter(prisma),
+  // 認証プロバイダーの設定
+  providers: [
+    // Google認証
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID as string,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+    }),
+    // メールアドレス認証
+    CredentialsProvider({
+      name: 'credentials',
+      credentials: {
+        // メールアドレスとパスワード
+        email: { label: 'email', type: 'text' },
+        password: { label: 'password', type: 'password' },
+      },
+      async authorize(credentials) {
+        // メールアドレスとパスワードがない場合はエラー
+        if (!credentials?.email || !credentials?.password) {
+          throw new Error('メールアドレスとパスワードが存在しません');
+        }
+
+        // ユーザーを取得
+        const user = await prisma.user.findUnique({
+          where: {
+            email: credentials.email,
+          },
+        });
+
+        // ユーザーが存在しない場合はエラー
+        if (!user || !user?.hashedPassword) {
+          throw new Error('ユーザーが存在しません');
+        }
+
+        // パスワードが一致しない場合はエラー
+        const isCorrectPassword = await bcrypt.compare(credentials.password, user.hashedPassword);
+
+        if (!isCorrectPassword) {
+          throw new Error('パスワードが一致しません');
+        }
+
+        return user;
+      },
+    }),
+  ],
+  // セッションの設定
+  session: {
+    strategy: 'jwt',
+  },
+};
+
+// 認証情報取得
+export const getAuthSession = async () => {
+  const session = await getServerSession(authOptions);
+
+  if (!session || !session.user?.email) {
+    return null;
+  }
+
+  // 1件のレコードを取得、見つからない場合は例外を投げる
+  const user = await prisma.user.findFirstOrThrow({
+    where: {
+      email: session.user.email,
+    },
+  });
+
+  return user;
+};
