--- frontend/src/shop/page.tsx
@@ -1,11 +1,0 @@
-import LoginForm from '@/components/LoginForm';
-
-export default async function Shop() {
-  return (
-    <main className="flex items-center justify-center md:h-screen">
-      <div className="relative mx-auto flex w-full max-w-[400px] flex-col space-y-2.5 p-4 md:-mt-32">
-        <LoginForm />
-      </div>
-    </main>
-  );
-}
